+++ src/main/java/com/chip8/api/core/memory/Memory.java
@@ -2,8 +2,8 @@ package com.chip8.api.core.memory;
 
 public interface Memory {
 
-    void write(Integer position, Byte data);
+    void write(Integer position, Integer data);
 
-    Byte read(Integer position);
+    Integer read(Integer position);
 
 }
