+++ src/main/java/com/chip8/api/core/register/IRegister.java
@@ -2,8 +2,8 @@ package com.chip8.api.core.register;
 
 public interface IRegister {
 
-    void setVI(Short data);
+    void setVI(Integer data);
 
-    Short getVI();
+    Integer getVI();
 
 }
