+++ src/main/java/com/chip8/model/core/register/IRegisterHandler.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Component;
 @Setter
 public class IRegisterHandler implements IRegister {
 
-    private Short vI;
+    private Integer vI;
 
 }
