+++ src/main/java/com/chip8/model/core/register/VectorRegisterHandler.java
@@ -7,10 +7,10 @@ public class VectorRegisterHandler implements VectorRegister {
 
     private static final Integer MAX_LENGTH_V_REGISTER = 16;
 
-    private final Byte[] v = new Byte[MAX_LENGTH_V_REGISTER];
+    private final Integer[] v = new Integer[MAX_LENGTH_V_REGISTER];
 
     @Override
-    public void setVRegister(final Integer position, final Byte data) {
+    public void setVRegister(final Integer position, final Integer data) {
         Assert.notNull(position, "The position can not be null");
         Assert.isTrue(position >= 0, "The position can not be negative");
         Assert.isTrue(position < this.v.length, "The position can not be greater than " + (this.v.length - 1));
@@ -18,7 +18,7 @@ public class VectorRegisterHandler implements VectorRegister {
     }
 
     @Override
-    public Byte getVRegister(final Integer position) {
+    public Integer getVRegister(final Integer position) {
         Assert.notNull(position, "The position can not be null");
         Assert.isTrue(position >= 0, "The position can not be negative");
         Assert.isTrue(position < this.v.length, "The position can not be greater than " + (this.v.length - 1));
