+++ src/test/java/com/chip8/model/core/instruction/CallAddressInstructionTest.java
@@ -50,14 +50,14 @@ class CallAddressInstructionTest {
 
     @Test
     void run_test() {
-        Mockito.when(this.sp.getSp()).thenReturn((short) 1);
-        Mockito.when(this.pc.getPc()).thenReturn((short) 2);
+        Mockito.when(this.sp.getSp()).thenReturn(1);
+        Mockito.when(this.pc.getPc()).thenReturn(2);
 
         this.callAddressInstruction.run("2FFF");
 
-        Mockito.verify(this.memoryStack, Mockito.times(1)).write(1, (byte) 2);
+        Mockito.verify(this.memoryStack, Mockito.times(1)).write(1, 2);
         Mockito.verify(this.sp, Mockito.times(1)).increase();
-        Mockito.verify(this.pc, Mockito.times(1)).setPc((short) HexFormat.fromHexDigits("2FFF".substring(1)));
+        Mockito.verify(this.pc, Mockito.times(1)).setPc(HexFormat.fromHexDigits("2FFF".substring(1)));
     }
 
 }
