+++ src/test/java/com/chip8/model/core/register/StackPointerHandlerTest.java
@@ -9,28 +9,28 @@ class StackPointerHandlerTest {
 
     @Test
     void decrement_stackPointerIsZero_resultNoDecrement_test() {
-        this.stackPointerHandler.setSp((short) 0);
+        this.stackPointerHandler.setSp(0);
         this.stackPointerHandler.decrement();
         Assertions.assertEquals((short) 0, this.stackPointerHandler.getSp());
     }
 
     @Test
     void decrement_stackPointerGreaterThanZero_resultDecrement_test() {
-        this.stackPointerHandler.setSp((short) 1);
+        this.stackPointerHandler.setSp(1);
         this.stackPointerHandler.decrement();
         Assertions.assertEquals((short) 0, this.stackPointerHandler.getSp());
     }
 
     @Test
     void increase_stackPointerIsZero_resultNoDecrement_test() {
-        this.stackPointerHandler.setSp((short) 11);
+        this.stackPointerHandler.setSp(11);
         this.stackPointerHandler.increase();
         Assertions.assertEquals((short) 11, this.stackPointerHandler.getSp());
     }
 
     @Test
     void increase_stackPointerGreaterThanZero_resultDecrement_test() {
-        this.stackPointerHandler.setSp((short) 10);
+        this.stackPointerHandler.setSp(10);
         this.stackPointerHandler.increase();
         Assertions.assertEquals((short) 11, this.stackPointerHandler.getSp());
     }
